#238 * Initial creation/migraton of mobile notification function to plugins.
